✨ update endpoints

+++ backend/controllers/product.controller.js
@@ -29,13 +29,17 @@ export const createProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
-  console.log("id", id);
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({ message: "Invalid product ID" });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ message: "Product deleted successfully" });
   } catch (error) {
     console.error(error);
-    res.status(404).json({ message: "Product not found" });
+    res.status(500).json({ message: "Product not found" });
   }
 };
